invoice methods and other layers

+++ src/main/java/com/inventory/Inventory/exception/ProductsNotFoundException.java
@@ -2,7 +2,7 @@ package com.inventory.Inventory.exception;
 
 public class ProductsNotFoundException extends RuntimeException{
 
-    public ProductsNotFoundException(Integer id_product){
+    public ProductsNotFoundException(Long id_product){
         super("Could not found the Product with id " + id_product);
     }
 }
