+++ src/main/java/com/inventory/Inventory/repository/ProductsRepository.java
@@ -3,5 +3,5 @@ package com.inventory.Inventory.repository;
 import com.inventory.Inventory.model.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Products, Integer> {
+public interface ProductsRepository extends JpaRepository<Products, Long> {
 }
